requirements.txt
tgcrypto
asyncio
ffmpeg-python
py-tgcalls==0.8.6
pyrogram==1.4.16
git+https://github.com/S780821/yt-dlp@master
youtube-search-python
youtube-search
python-dotenv
dnspython
gitpython
aiofiles
aiohttp
requests
pillow
heroku3
motor
psutil
future
wget
pyyaml==5.3.1
telethon
speedtest-cli
lyricsgenius
